+++ src/Service/ExampleAction.php
@@ -4,11 +4,17 @@
 namespace Tzm\Authorizator\Service;
 
 
+use Tzm\Authorizator\Service\AuthorizatorChannels\ExampleChannel;
+
 class ExampleAction extends AuthorizatorAction
 {
 
+    protected $allowedChannels = [
+        ExampleChannel::class,
+    ];
+
     public function afterAuthorization()
     {
-        // TODO: Implement afterAuthorization() method.
+        session(['afterAuthorization' => true]);
     }
 }