+++ src/tests/Feature/ActionTests.php
@@ -2,11 +2,7 @@
 
 namespace Tzm\Authorizator\Feature;
 
-use App\User;
-use Illuminate\Foundation\Testing\DatabaseMigrations;
-use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tzm\Authorizator\AuthorizatorTestCase;
-use Tzm\Authorizator\Model\Authorization;
 
 class ActionTests extends AuthorizatorTestCase
 {
@@ -14,12 +10,9 @@ class ActionTests extends AuthorizatorTestCase
     /** @test */
     public function sendCode()
     {
-        $user = factory(User::class)->create();
-        $posts = $this->factory->of(Authorization::class)->create();
+        $response = $this->get('/');
 
-        $this->actingAs($user);
-        $this->post(route('authorizator.check'))->assertStatus(200);
-
+        $response->assertStatus(200);
     }
 
     /** @test */
@@ -43,7 +36,6 @@ class ActionTests extends AuthorizatorTestCase
     {
 
     }
-
     /** @test */
     public function staticDeliverCodeToUser()
     {